# Convert vector to f16 for dequantize mul mat vec