Currently dequantize mul mat vec operates on a quantized matrix and an f32 vector. This PR converts the vector to f16 and then converts it back to f32 as it's needed. Because this way each streaming multiprocessor needs to load less data during the calculation this is actually slightly faster:

| GPU      | Test   | Model    | t/s master | t/s PR | Speedup |
|----------|--------|----------|------------|--------|---------|
| RTX 3090 | TG 128 | 7b q4_0  |      83.26 |  90.09 |    1.08 |
| RTX 3090 | TG 128 | 13b q4_0 |      46.08 |  50.56 |    1.10 |
| RTX 3090 | TG 128 | 33b q4_0 |      20.21 |  22.24 |    1.10 |
| GTX 1070 | TG 128 | 7b q4_0  |      15.93 |  16.25 |    1.02 |

I also tried a version in which the vector is instead converted to q8_0 but I was not able to get performance out of this; int <-> float conversion and integer arithmetic may be too slow on GPUs to make this worthwhile. There is also a [version which uses f16 arithmetic](https://github.com/JohannesGaessler/llama.cpp/tree/cuda-dmmv-f16) but that version was only ~3% faster on an RTX 3090 and would require relatively large changes to the code. The performance on my GTX 1070 was also very bad because Pascal seems to have poor f16 performance.

I did not touch any of the k-quants because they use the vector quite differently than the older quantization methods and I wasn't sure what the correct implementation would be.