@ikawrakow 2023-06-17
The change looks good, but in my case the PR leads to a lower performance on my RTX-4080. I get 110 t/s on master (b2416493ab3ab21686d47c96669da6d6c6af08a4) for `Q4_0` at 7B, but 98 t/s on this branch, with and without rebasing on current master. It is late here, so maybe I'm making a stupid mistake. I'll check again tomorrow.