@JohannesGaessler 2023-06-18
Already on it, I've already added an option `LLAMA_CUDA_DMMV_F16` to the Makefile and put more effort in the implementation. The new implementation uses f16 intrinsics if the option is enabled and I optimized the data types in the DMMV code in general so it should now be very slightly faster:

| GPU      | Test   | Model    | t/s master | t/s PR | t/s PR LLAMA_CUDA_DMMV_F16=1 | Speedup PR LLAMA_CUDA_DMMV_F16=1 vs. master |
|----------+--------+----------+------------+--------+------------------------------+---------------------------------------------|
| RTX 3090 | TG 128 | 7b q4_0  |      84.60 |  85.18 |                        91.49 |  1.08                                       |
| RTX 3090 | TG 128 | 13b q4_0 |      46.49 |  46.21 |                        51.44 |  1.11                                       |
| RTX 3090 | TG 128 | 33b q4_0 |      20.12 |  20.05 |                        22.64 |  1.13                                       |
| RTX 3090 | TG 128 | 7b q4_1  |      44.68 |  45.31 |                        48.70 |  1.09                                       |
| RTX 3090 | TG 128 | 7b q5_0  |      37.17 |  37.51 |                        35.06 |  0.94                                       |
| RTX 3090 | TG 128 | 7b q5_1  |      36.54 |  36.56 |                        35.10 |  0.96                                       |
| RTX 3090 | TG 128 | 7b q8_0  |      68.02 |  68.11 |                        73.64 |  1.08                                       |
| RTX 3090 | TG 128 | 7b f16   |      51.49 |  51.31 |                        50.48 |  0.98                                       |

I still need to do more testing though.