@JohannesGaessler 2023-06-18
Some more testing:

| GPU                    | Test   | Model   | t/s master | t/s PR | t/s PR LLAMA_CUDA_DMMV_F16=1 |
|------------------------|--------|---------|------------|--------|------------------------------|
| GTX 1070               | TG 128 | 7b q4_0 |      16.79 |  17.08 | 2.75                         |
| GTX 1070               | TG 128 | 7b q4_1 |      14.24 |  14.20 | Not tested                   |
| GTX 1070               | TG 128 | 7b q5_0 |       9.56 |   9.57 | Not tested                   |
| GTX 1070               | TG 128 | 7b q5_1 |       9.33 |   9.31 | Not tested                   |
| GTX 1070 + GTX 1050 ti | TG 128 | 7b q4_0 |      19.50 |  19.77 | 3.62                         |
| GTX 1070 + GTX 1050 ti | TG 128 | 7b q4_1 |      17.02 |  17.02 | Not tested                   |
| GTX 1070 + GTX 1050 ti | TG 128 | 7b q5_0 |      12.04 |  12.03 | Not tested                   |
| GTX 1070 + GTX 1050 ti | TG 128 | 7b q5_1 |      11.73 |  11.74 | Not tested                   |
| GTX 1070 + GTX 1050 ti | TG 128 | 7b q8_0 |      14.00 |  14.00 | Not tested                   |

As expected the f16 performance on my Pascal cards is very bad.